Fix regression in log auto-updating code
Also switch back to using integers for the slider min and max
values, and for the hc.front.forms.LogFilterForm.

+++ hc/front/forms.py
@@ -383,7 +383,7 @@ class SearchForm(forms.Form):
 
 class LogFiltersForm(forms.Form):
     # min_value is 2010-01-01, max_value is 2030-01-01
-    end = forms.FloatField(min_value=1262296800, max_value=1893448800, required=False)
+    end = forms.IntegerField(min_value=1262296800, max_value=1893448800, required=False)
 
     def clean_end(self) -> datetime | None:
         if self.cleaned_data["end"]:
